Move rbac package under /pkg
